generate-provision-quality cli command
