+++ digital_land/cli.py
@@ -31,6 +31,7 @@ from digital_land.commands import (
     organisation_check,
     save_state,
     add_data,
+    generate_provision_quality,
 )
 
 from digital_land.command_arguments import (
@@ -825,3 +826,8 @@ def check_state_cmd(
     if diffs:
         print(f"State differs from {state_path} - {', '.join(diffs)}")
         sys.exit(1)
+
+
+@cli.command("generate-provision-quality")
+def generate_provision_quality_cmd():
+    generate_provision_quality()
