+++ setup.py
@@ -61,6 +61,7 @@ setup(
         "boto3",
         "moto",
         "psutil",
+        "geopandas",
     ],
     entry_points={"console_scripts": ["digital-land=digital_land.cli:cli"]},
     setup_requires=["pytest-runner"],
